Split RelationshipValidator and generating missing relationships
