+++ estep/validate.py
@@ -20,6 +20,8 @@ import logging
 import jsonschema
 import requests
 import six
+from . import relationship
+from .relationship import AbstractValidator
 
 try:
     from jsonschema._format import is_uri as is_uri_orig
@@ -29,6 +31,19 @@ except ImportError:
 
 LOGGER = logging.getLogger('estep')
 
+
+def url_to_path(url):
+    prefix, path = url.split('://software.esciencecenter.nl/')
+    if prefix == 'https':
+        raise ValueError('For the time being, use http instead of https prefixes for http://software.esciencecenter.nl')
+
+    # remove additional indicator
+    path = path.split('#')[0]
+    return '_' + path + '.md'
+
+def url_to_collection_name(url):
+    path = url.split('://software.esciencecenter.nl/')[1]
+    return path.split('/')[0]
 
 def url_ref(instance):
     return url_local_ref(instance) and url_resolve(instance)
@@ -40,13 +55,7 @@ def url_local_ref(instance):
 
     # lookup local files locally
     if '://software.esciencecenter.nl/' in instance:
-        prefix, path = instance.split('://software.esciencecenter.nl/')
-        if prefix == 'https':
-            raise ValueError('For the time being, use http instead of https prefixes for http://software.esciencecenter.nl')
-
-        # remove additional indicator
-        path = path.split('#')[0]
-        location = '_' + path + '.md'
+        location = url_to_path(instance)
         # do not look for missing directories.
         if os.path.isdir(os.path.dirname(location)) and not os.path.isfile(location):
             err = "{} not found locally as {}".format(instance, location)
@@ -76,16 +85,6 @@ def log_error(error, prefix=""):
     LOGGER.warning(msg)
 
 
-class AbstractValidator(object):
-    def validate(self, name, instance):
-        # TODO yield errors instead of logging errors and returning nr_errors
-        raise NotImplementedError
-
-    def finalize(self):
-        # TODO yield errors instead of logging errors and returning nr_errors
-        return 0
-
-
 class AbstractValidators(AbstractValidator, six.moves.UserList):
     def validate(self, name, instance):
         nr_errors = 0
@@ -99,6 +98,10 @@ class AbstractValidators(AbstractValidator, six.moves.UserList):
             nr_errors += validator.finalize()
         return nr_errors
 
+    def missing(self):
+        for validator in self.data:
+            for missing_tuple in validator.missing():
+                yield missing_tuple
 
 class PropertyTypoValidator(AbstractValidator):
     """
@@ -137,169 +140,31 @@ class PropertyTypoValidator(AbstractValidator):
         return nr_errors
 
 
-class RelationShipValidator(AbstractValidator):
-    """
+def load_schemas(schema_uris, schemadir=None):
+    store = {}
+    for schema_uri in schema_uris:
+        if schemadir is None:
+            u = schema_uri
+            request = requests.get(u)
+            # do not accept failed calls
+            try:
+                request.raise_for_status()
+            except requests.exceptions.HTTPError as ex:
+                LOGGER.error("cannot load schema %s:\n\t%s\nUse --local to load local schemas.", schema_uri, ex)
 
-    >>> validator = RelationShipValidator('http://software.esciencecenter.nl/schema/software', 'usedIn',
-    ...                                   'http://software.esciencecenter.nl/schema/project', 'uses')
-    >>> validator.validate('twinl', {'schema': 'http://software.esciencecenter.nl/schema/project',
-    ...                              '@id': 'http://software.esciencecenter.nl/project/twinl',
-    ...                              'uses': ['http://software.esciencecenter.nl/software/twiqs.nl']})
-    ... # twinl uses twiqs.nl, but twiqs.nl has not been seen yet, so OK for now
-    0
-    >>> validator.validate('twiqs.nl', {'schema': 'http://software.esciencecenter.nl/schema/software',
-    ...                                 '@id': 'http://software.esciencecenter.nl/software/twiqs.nl',
-    ...                                 'usedIn': ['http://software.esciencecenter.nl/project/twinl']})
-    ... # twiqs.nl is used in twinl and twinl uses twiqs.nl, so is OK
-    0
-    >>> validator.validate('emetabolomics', {'schema': 'http://software.esciencecenter.nl/schema/project',
-    ...                                      '@id': 'http://software.esciencecenter.nl/project/emetabolomics',
-    ...                                      'uses': ['http://software.esciencecenter.nl/software/xenon']})
-    ... # xenon is used by emetabolomics, but xenon has not been seen yet, so OK for now
-    0
-    >>> validator.validate('xenon', {'schema': 'http://software.esciencecenter.nl/schema/software',
-    ...                              '@id': 'http://software.esciencecenter.nl/software/xenon',
-    ...                              'usedIn': ['http://software.esciencecenter.nl/project/simcity']})
-    ... # xenon is used by emetabolomics, but xenon is not used in emetabolomics, so gives error
-    1
-    >>> validator.validate('simcity', {'schema': 'http://software.esciencecenter.nl/schema/project',
-    ...                                '@id': 'http://software.esciencecenter.nl/project/simcity',
-    ...                                'uses': ['http://software.esciencecenter.nl/software/pyxenon']})
-    ... # xenon is used in simcity, but simcity does not use xenon, so gives error
-    1
-    >>> validator.finalize()
-    ... # simcity uses pyxenon, but validator has not seen pyxenon, so gives error
-    1
+            store[schema_uri] = request.json()
+        else:
+            LOGGER.debug('Loading schema %s from %s', schema_uri, schemadir)
+            schema_fn = schema_uri.replace('http://software.esciencecenter.nl/schema', schemadir)
+            with open(schema_fn) as f:
+                store[schema_uri] = json.load(f)
+    return store
 
-    """
-    def __init__(self, schema1, prop1, schema2, prop2):
-        self.schema1 = schema1
-        self.prop1 = prop1
-        self.memory1 = {}
-        self.schema2 = schema2
-        self.prop2 = prop2
-        self.memory2 = {}
-        self.errors = set()
-
-    def validate(self, name, instance):
-        schema = instance['schema']
-        myid = instance['@id']
-        nr_errors = 0
-
-        if schema == self.schema1 and self.prop1 in instance:
-            if isinstance(instance[self.prop1], str):
-                myvalues = set([instance[self.prop1]])
-            elif isinstance(instance[self.prop1], dict):
-                # not a @id
-                return nr_errors
-            else:
-                myvalues = set([d for d in instance[self.prop1] if isinstance(d, str)])
-
-            for myvalue in myvalues:
-                if myvalue not in self.memory2:
-                    LOGGER.debug('Unable to validate {0} {1} {2}, have not seen counterpart yet'.format(myid, self.prop1, myvalue))
-                    continue
-                if myid in self.memory2[myvalue]:
-                    LOGGER.debug('{0} {1} {2} and {2} {3} {0}'.format(myid, self.prop1, myvalue, self.prop2))
-                    continue
-                error = 'Missing "{0}"'.format(myid)
-                LOGGER.warning('* Error      : ' + error)
-                LOGGER.warning('  On property: ' + self.prop2 + ' of ' + myvalue)
-                self.errors.add((myid, self.prop2, myvalue))
-                nr_errors += 1
-
-            # find other direction
-            for otherid, othervalues in six.iteritems(self.memory2):
-                if myid in othervalues and otherid not in myvalues:
-                    error = 'Missing "{0}"'.format(otherid)
-                    LOGGER.warning('* Error      : ' + error)
-                    LOGGER.warning('  On property: ' + self.prop1)
-                    self.errors.add((otherid, self.prop1, myid))
-                    nr_errors += 1
-
-            self.memory1[myid] = myvalues
-        if schema == self.schema2 and self.prop2 in instance:
-            if isinstance(instance[self.prop2], str):
-                myvalues = set([instance[self.prop2]])
-            elif isinstance(instance[self.prop2], dict):
-                # not a @id
-                return nr_errors
-            else:
-                myvalues = set([d for d in instance[self.prop2] if isinstance(d, str)])
-
-            for myvalue in myvalues:
-                if myvalue not in self.memory1:
-                    LOGGER.debug('Unable to validate {0} {1} {2}, have not seen counterpart yet'.format(myid, self.prop2, myvalue))
-                    continue
-                if myid in self.memory1[myvalue]:
-                    LOGGER.debug('{0} {1} {2} and {2} {3} {0}'.format(myid, self.prop2, myvalue, self.prop1))
-                    continue
-                error = 'Missing "{0}"'.format(myid)
-                LOGGER.warning('* Error      : ' + error)
-                LOGGER.warning('  On property: ' + self.prop1 + ' of ' + myvalue)
-                self.errors.add((myid, self.prop1, myvalue))
-                nr_errors += 1
-
-            # find other direction
-            for otherid, othervalues in six.iteritems(self.memory1):
-                if myid in othervalues and otherid not in myvalues:
-                    error = 'Missing "{0}"'.format(otherid)
-                    LOGGER.warning('* Error      : ' + error)
-                    LOGGER.warning('  On property: ' + self.prop2)
-                    self.errors.add((otherid, self.prop2, myid))
-                    nr_errors += 1
-
-            self.memory2[myid] = myvalues
-
-        return nr_errors
-
-    def finalize(self):
-        nr_errors = 0
-        for myid, myvalues in six.iteritems(self.memory1):
-            for myvalue in myvalues:
-                if myvalue in self.memory2 and myid in self.memory2[myvalue]:
-                    pass
-                else:
-                    error = 'Missing "{0}"'.format(myid)
-                    if (myid, self.prop2, myvalue) not in self.errors:
-                        LOGGER.warning('* Error      : ' + error)
-                        LOGGER.warning('  On property: ' + self.prop2 + ' of ' + myvalue)
-                        nr_errors += 1
-
-        for myid, myvalues in six.iteritems(self.memory2):
-            for myvalue in myvalues:
-                if myvalue in self.memory1 and myid in self.memory1[myvalue]:
-                    pass
-                else:
-                    # error = '"{0}" not found locally'.format(myvalue)
-                    error = 'Missing "{0}"'.format(myid)
-                    if (myid, self.prop1, myvalue) not in self.errors:
-                        LOGGER.warning('* Error      : ' + error)
-                        LOGGER.warning('  On property: ' + self.prop1 + ' of ' + myvalue)
-                        nr_errors += 1
-
-        return nr_errors
 
 
 class Validator(AbstractValidator):
     def __init__(self, schema_uris, schemadir=None, resolve_local=True, resolve_remote=False):
-        store = {}
-        for schema_uri in schema_uris:
-            if schemadir is None:
-                u = schema_uri
-                request = requests.get(u)
-                # do not accept failed calls
-                try:
-                    request.raise_for_status()
-                except requests.exceptions.HTTPError as ex:
-                    LOGGER.error("cannot load schema %s:\n\t%s\nUse --local to load local schemas.", schema_uri, ex)
-
-                store[schema_uri] = request.json()
-            else:
-                LOGGER.debug('Loading schema %s from %s', schema_uri, schemadir)
-                schema_fn = schema_uri.replace('http://software.esciencecenter.nl/schema', schemadir)
-                with open(schema_fn) as f:
-                    store[schema_uri] = json.load(f)
+        store = load_schemas(schema_uris, schemadir)
 
         # Resolve date-time as dates as well as strings
         if isinstance(jsonschema.compat.str_types, type):
@@ -333,62 +198,7 @@ class Validator(AbstractValidator):
         self.crossValidators.append(PropertyTypoValidator('programmingLanguage'))
         self.crossValidators.append(PropertyTypoValidator('technologyTag'))
         # From software
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'usedIn',
-                                                          'http://software.esciencecenter.nl/schema/project',
-                                                          'uses'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'dependency',
-                                                          'http://software.esciencecenter.nl/schema/software',
-                                                          'dependencyOf'))
-        # FIXME person:contactPerson is for project only? so below not OK?
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'contactPerson',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'contactPersonOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'owner',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'ownerOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'owner',
-                                                          'http://software.esciencecenter.nl/schema/organization',
-                                                          'ownerOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'user',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'userOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'user',
-                                                          'http://software.esciencecenter.nl/schema/organization',
-                                                          'userOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/software',
-                                                          'contributor',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'contributorOf'))
-        # FIXME software:involvedOrganization has no counterpart in organization schema
-        # From project
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/project',
-                                                          'contactPerson',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'contactPersonOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/project',
-                                                          'coordinator',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'coordinatorOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/project',
-                                                          'engineer',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'engineerOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/project',
-                                                          'principalInvestigator',
-                                                          'http://software.esciencecenter.nl/schema/person',
-                                                          'principalInvestigatorOf'))
-        self.crossValidators.append(RelationShipValidator('http://software.esciencecenter.nl/schema/project',
-                                                          'involvedOrganization',
-                                                          'http://software.esciencecenter.nl/schema/organization',
-                                                          'involvedIn'))
-        # FIXME person:affiliation has no counterpart in organization schema
+        self.crossValidators += relationship.get_validators()
 
     def validate(self, name, instance):
         schema_uri = instance['schema']
